+++ src/gmail/util.ts
@@ -1,28 +1,28 @@
 import { debugGmail } from '../debuggers';
 import Accounts, { IAccount } from '../models/Accounts';
-import { getAuth, gmailClient } from './auth';
+import { getCommonGoogleConfigs, getConfig } from '../utils';
+import { getOauthClient, gmailClient } from './auth';
 import { ICredentials } from './types';
 
 /**
  * Gets the current user's Gmail profile
  */
 export const getProfile = async (credentials: ICredentials, email?: string) => {
-  const auth = getAuth(credentials);
-
   debugGmail(`Gmail get an user profile`);
 
-  let userProfile;
-
   try {
-    userProfile = await gmailClient.getProfile({
+    const auth = await getOauthClient();
+
+    auth.setCredentials(credentials);
+
+    return gmailClient.getProfile({
       auth,
       userId: email || 'me',
     });
   } catch (e) {
     debugGmail(`Error Google: Gmail failed to get user profile ${e}`);
+    throw e;
   }
-
-  return userProfile;
 };
 
 export const getCredentialsByEmailAccountId = async ({
@@ -58,7 +58,7 @@ export const getCredentialsByEmailAccountId = async ({
 export const getCredentials = (credentials: IAccount): ICredentials => ({
   access_token: credentials.token,
   refresh_token: credentials.tokenSecret,
-  expiry_date: credentials.expireDate,
+  expiry_date: parseInt(credentials.expireDate, 10),
   scope: credentials.scope,
 });
 
@@ -225,4 +225,22 @@ export const buildEmail = (rawString: string) => {
       }
     })
     .filter(email => email !== undefined);
+};
+
+export const getGoogleConfigs = async () => {
+  const {
+    GOOGLE_CLIENT_ID,
+    GOOGLE_CLIENT_SECRET,
+    GOOGLE_PROJECT_ID,
+    GOOGLE_APPLICATION_CREDENTIALS,
+  } = await getCommonGoogleConfigs();
+
+  return {
+    GOOGLE_PROJECT_ID,
+    GOOGLE_CLIENT_ID,
+    GOOGLE_CLIENT_SECRET,
+    GOOGLE_GMAIL_TOPIC: await getConfig('GOOGLE_GMAIL_TOPIC', 'gmail_topic'),
+    GOOGLE_APPLICATION_CREDENTIALS,
+    GOOGLE_GMAIL_SUBSCRIPTION_NAME: await getConfig('GOOGLE_GMAIL_SUBSCRIPTION_NAME', 'gmail_topic_subscription'),
+  };
 };