+++ src/helpers.ts
@@ -4,7 +4,7 @@ import {
   Conversations as ChatfuelConversations,
   Customers as ChatfuelCustomers,
 } from './chatfuel/models';
-import { debugCallPro, debugFacebook, debugGmail, debugNylas, debugTwitter } from './debuggers';
+import { debugCallPro, debugFacebook, debugGmail, debugNylas, debugTwitter, debugWhatsapp } from './debuggers';
 import {
   Comments as FacebookComments,
   ConversationMessages as FacebookConversationMessages,
@@ -18,10 +18,11 @@ import {
   Conversations as GmailConversations,
   Customers as GmailCustomers,
 } from './gmail/models';
-import { getCredentialsByEmailAccountId } from './gmail/util';
 import { stopPushNotification } from './gmail/watch';
 import { Accounts, Integrations } from './models';
-import { enableOrDisableAccount } from './nylas/auth';
+import Configs from './models/Configs';
+import { enableOrDisableAccount, removeExistingNylasWebhook } from './nylas/auth';
+import { setupNylas } from './nylas/controller';
 import {
   NylasGmailConversationMessages,
   NylasGmailConversations,
@@ -39,28 +40,35 @@ import {
   NylasYahooConversations,
   NylasYahooCustomers,
 } from './nylas/models';
-import { unsubscribe } from './twitter/api';
+import { createNylasWebhook } from './nylas/tracker';
+import { getTwitterConfig, unsubscribe } from './twitter/api';
+import * as twitterApi from './twitter/api';
 import {
   ConversationMessages as TwitterConversationMessages,
   Conversations as TwitterConversations,
   Customers as TwitterCustomers,
 } from './twitter/models';
-import { getEnv, sendRequest } from './utils';
+import { getEnv, resetConfigsCache, sendRequest } from './utils';
+import {
+  ConversationMessages as WhatsappConversationMessages,
+  Conversations as WhatsappConversations,
+  Customers as WhatsappCustomers,
+} from './whatsapp/models';
 
-/**
- * Remove integration integrationId
- */
+import { logout, setupChatApi as setupWhatsapp } from './whatsapp/api';
+
 export const removeIntegration = async (integrationErxesApiId: string): Promise<string> => {
   const integration = await Integrations.findOne({ erxesApiId: integrationErxesApiId });
 
   if (!integration) {
-    return;
+    throw new Error('Integration not found');
   }
 
   // Remove endpoint
   let integrationRemoveBy;
 
   const { _id, kind, accountId, erxesApiId } = integration;
+
   const account = await Accounts.findOne({ _id: accountId });
 
   const selector = { integrationId: _id };
@@ -75,11 +83,18 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
         pageTokenResponse = await getPageAccessToken(pageId, account.token);
       } catch (e) {
         debugFacebook(`Error ocurred while trying to get page access token with ${e.message}`);
+        throw e;
       }
 
       await FacebookPosts.deleteMany({ recipientId: pageId });
       await FacebookComments.deleteMany({ recipientId: pageId });
-      await unsubscribePage(pageId, pageTokenResponse);
+
+      try {
+        await unsubscribePage(pageId, pageTokenResponse);
+      } catch (e) {
+        debugFacebook(`Error occured while trying to unsubscribe page pageId: ${pageId}`);
+        throw e;
+      }
     }
 
     integrationRemoveBy = { fbPageIds: integration.facebookPageIds };
@@ -96,12 +111,16 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
   if (kind === 'gmail' && !account.nylasToken) {
     debugGmail('Removing gmail entries');
 
-    const credentials = await getCredentialsByEmailAccountId({ email: account.uid });
     const conversationIds = await GmailConversations.find(selector).distinct('_id');
 
     integrationRemoveBy = { email: integration.email };
 
-    await stopPushNotification(account.uid, credentials);
+    try {
+      await stopPushNotification(account.uid);
+    } catch (e) {
+      debugGmail('Failed to stop push notification of gmail account');
+      throw e;
+    }
 
     await GmailCustomers.deleteMany(selector);
     await GmailConversations.deleteMany(selector);
@@ -117,8 +136,13 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
     await NylasGmailConversations.deleteMany(selector);
     await NylasGmailConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
 
-    // Cancel nylas subscription
-    await enableOrDisableAccount(account.uid, false);
+    try {
+      // Cancel nylas subscription
+      await enableOrDisableAccount(account.uid, false);
+    } catch (e) {
+      debugNylas('Failed to cancel nylas-gmail account subscription');
+      throw e;
+    }
   }
 
   if (kind === 'callpro') {
@@ -137,11 +161,33 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
 
     const conversationIds = await TwitterConversations.find(selector).distinct('_id');
 
-    unsubscribe(account.uid);
+    try {
+      unsubscribe(account.uid);
+    } catch (e) {
+      debugNylas('Failed to unsubscribe twitter account');
+      throw e;
+    }
 
     await TwitterConversationMessages.deleteMany(selector);
     await TwitterConversations.deleteMany(selector);
     await TwitterCustomers.deleteMany({ conversationId: { $in: conversationIds } });
+  }
+
+  if (kind === 'whatsapp') {
+    debugWhatsapp('Removing whatsapp entries');
+
+    try {
+      await logout(integration.whatsappinstanceId, integration.whatsappToken);
+    } catch (e) {
+      debugWhatsapp('Failed to logout WhatsApp account');
+      throw e;
+    }
+
+    const conversationIds = await WhatsappConversations.find(selector).distinct('_id');
+
+    await WhatsappConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
+    await WhatsappConversations.deleteMany(selector);
+    await WhatsappCustomers.deleteMany(selector);
   }
 
   // Remove from core =========
@@ -173,8 +219,13 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
     await NylasImapConversations.deleteMany(selector);
     await NylasImapConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
 
-    // Cancel nylas subscription
-    await enableOrDisableAccount(account.uid, false);
+    try {
+      // Cancel nylas subscription
+      await enableOrDisableAccount(account.uid, false);
+    } catch (e) {
+      debugNylas('Failed to cancel subscription of nylas-imap account');
+      throw e;
+    }
   }
 
   if (kind === 'office365') {
@@ -186,8 +237,13 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
     await NylasOffice365Conversations.deleteMany(selector);
     await NylasOffice365ConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
 
-    // Cancel nylas subscription
-    await enableOrDisableAccount(account.uid, false);
+    try {
+      // Cancel nylas subscription
+      await enableOrDisableAccount(account.uid, false);
+    } catch (e) {
+      debugNylas('Failed to subscription nylas-office365 account');
+      throw e;
+    }
   }
 
   if (kind === 'outlook') {
@@ -199,8 +255,13 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
     await NylasOutlookConversations.deleteMany(selector);
     await NylasOutlookConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
 
-    // Cancel nylas subscription
-    await enableOrDisableAccount(account.uid, false);
+    try {
+      // Cancel nylas subscription
+      await enableOrDisableAccount(account.uid, false);
+    } catch (e) {
+      debugNylas('Failed to subscription nylas-outlook account');
+      throw e;
+    }
   }
 
   if (kind === 'yahoo') {
@@ -212,8 +273,13 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
     await NylasYahooConversations.deleteMany(selector);
     await NylasYahooConversationMessages.deleteMany({ conversationId: { $in: conversationIds } });
 
-    // Cancel nylas subscription
-    await enableOrDisableAccount(account.uid, false);
+    try {
+      // Cancel nylas subscription
+      await enableOrDisableAccount(account.uid, false);
+    } catch (e) {
+      debugNylas('Failed to subscription nylas-yahoo account');
+      throw e;
+    }
   }
 
   if (kind === 'chatfuel') {
@@ -231,23 +297,120 @@ export const removeIntegration = async (integrationErxesApiId: string): Promise<
   return erxesApiId;
 };
 
-/**
- * Remove integration by or accountId
- */
-export const removeAccount = async (_id: string): Promise<{ erxesApiIds: string | string[] }> => {
+export const removeAccount = async (_id: string): Promise<{ erxesApiIds: string | string[] } | Error> => {
   const account = await Accounts.findOne({ _id });
 
   if (!account) {
-    return;
+    return new Error(`Account not found: ${_id}`);
   }
 
   const erxesApiIds = [];
 
   const integrations = await Integrations.find({ accountId: account._id });
 
-  for (const integration of integrations) {
-    erxesApiIds.push(await removeIntegration(integration.erxesApiId));
+  if (integrations.length) {
+    for (const integration of integrations) {
+      try {
+        const response = await removeIntegration(integration.erxesApiId);
+        erxesApiIds.push(response);
+      } catch (e) {
+        throw e;
+      }
+    }
   }
 
+  await Accounts.deleteOne({ _id });
+
   return { erxesApiIds };
+};
+
+export const removeCustomers = async params => {
+  const { customerIds } = params;
+  const selector = { erxesApiId: { $in: customerIds } };
+
+  await FacebookCustomers.deleteMany(selector);
+  await NylasGmailCustomers.deleteMany(selector);
+  await NylasOutlookCustomers.deleteMany(selector);
+  await NylasOffice365Customers.deleteMany(selector);
+  await NylasYahooCustomers.deleteMany(selector);
+  await NylasImapCustomers.deleteMany(selector);
+  await ChatfuelCustomers.deleteMany(selector);
+  await CallProCustomers.deleteMany(selector);
+  await TwitterCustomers.deleteMany(selector);
+  await WhatsappCustomers.deleteMany(selector);
+};
+
+export const updateIntegrationConfigs = async (configsMap): Promise<void> => {
+  const getValueAsString = async name => {
+    const entry = await Configs.getConfig(name);
+
+    if (entry.value) {
+      return entry.value.toString();
+    }
+
+    return entry.value;
+  };
+
+  const prevNylasClientId = await getValueAsString('NYLAS_CLIENT_ID');
+  const prevNylasClientSecret = await getValueAsString('NYLAS_CLIENT_SECRET');
+  const prevNylasWebhook = await getValueAsString('NYLAS_WEBHOOK_CALLBACK_URL');
+
+  const prevChatApiWebhook = await getValueAsString('CHAT_API_WEBHOOK_CALLBACK_URL');
+  const prevChatApiUID = await getValueAsString('CHAT_API_UID');
+  const prevTwitterConfig = await getTwitterConfig();
+
+  await Configs.updateConfigs(configsMap);
+
+  const updatedTwitterConfig = await getTwitterConfig();
+
+  resetConfigsCache();
+
+  const updatedNylasClientId = await getValueAsString('NYLAS_CLIENT_ID');
+  const updatedNylasClientSecret = await getValueAsString('NYLAS_CLIENT_SECRET');
+  const updatedNylasWebhook = await getValueAsString('NYLAS_WEBHOOK_CALLBACK_URL');
+
+  const updatedChatApiWebhook = await getValueAsString('CHAT_API_WEBHOOK_CALLBACK_URL');
+  const updatedChatApiUID = await getValueAsString('CHAT_API_UID');
+
+  try {
+    if (prevNylasClientId !== updatedNylasClientId || prevNylasClientSecret !== updatedNylasClientSecret) {
+      await setupNylas();
+
+      await removeExistingNylasWebhook();
+      await createNylasWebhook();
+    }
+
+    if (prevNylasWebhook !== updatedNylasWebhook) {
+      await removeExistingNylasWebhook();
+      await createNylasWebhook();
+    }
+  } catch (e) {
+    debugNylas(e);
+  }
+
+  try {
+    if (
+      prevTwitterConfig.oauth.consumer_key !== updatedTwitterConfig.oauth.consumer_key ||
+      prevTwitterConfig.oauth.consumer_secret !== updatedTwitterConfig.oauth.consumer_secret
+    ) {
+      await twitterApi.registerWebhook();
+    }
+    if (
+      prevTwitterConfig.oauth.token !== updatedTwitterConfig.oauth.token ||
+      prevTwitterConfig.oauth.token_secret !== prevTwitterConfig.oauth.token_secret ||
+      prevTwitterConfig.twitterWebhookEnvironment !== updatedTwitterConfig.twitterWebhookEnvironment
+    ) {
+      await twitterApi.registerWebhook();
+    }
+  } catch (e) {
+    debugTwitter(e);
+  }
+
+  if (prevChatApiWebhook !== updatedChatApiWebhook || prevChatApiUID !== updatedChatApiUID) {
+    try {
+      await setupWhatsapp();
+    } catch (e) {
+      debugWhatsapp(e);
+    }
+  }
 };