+++ src/models/Integrations.ts
@@ -9,10 +9,12 @@ export interface IIntegration {
   facebookPageTokensMap?: { [key: string]: string };
   email: string;
   phoneNumber: string;
-  tenant: string;
+  recordUrl: string;
   expiration?: string;
   gmailHistoryId?: string;
   chatfuelConfigs?: { [key: string]: string };
+  whatsappinstanceId?: string;
+  whatsappToken?: string;
 }
 
 export interface IIntegrationDocument extends IIntegration, Document {}
@@ -24,7 +26,7 @@ export const integrationSchema = new Schema({
   accountId: String,
   erxesApiId: String,
   phoneNumber: String,
-  tenant: String,
+  recordUrl: String,
   facebookPageIds: [String],
   email: String,
   expiration: String,
@@ -37,6 +39,8 @@ export const integrationSchema = new Schema({
     type: Object,
     default: {},
   }),
+  whatsappinstanceId: String,
+  whatsappToken: String,
 });
 
 export interface IIntegrationModel extends Model<IIntegrationDocument> {